% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_barplot.R
\name{create_barplot}
\alias{create_barplot}
\title{Create Barplots}
\usage{
create_barplot(
  dataset,
  variable,
  treatment = "Treatment",
  strata = NULL,
  xlab = treatment,
  ylab = variable
)
}
\arguments{
\item{dataset}{Data.frame of analytic data}

\item{variable}{Outcome variable to plot}

\item{treatment}{Treatment variable name to summarize \code{variable}  by}

\item{strata}{Option variable to stratify figure by}

\item{xlab}{Character string to label x-axis. Defaults to \code{treatment}}

\item{ylab}{Character string to label y-axis. Defaults to \code{variable}}
}
\description{
plot: barplot using raw data: mean +/- SE
}
