% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create Boxplots with Significance Annotations}
\usage{
create_boxplot(
  dataset,
  variable,
  treatment = "Treatment",
  strata = NULL,
  contrast_p = NULL,
  xlab = treatment,
  ylab = variable
)
}
\arguments{
\item{dataset}{Data.frame of analytic data}

\item{variable}{Outcome variable to plot}

\item{treatment}{Treatment variable name to summarize \code{variable} by}

\item{strata}{Optional variable to stratify figure by}

\item{contrast_p}{Data frame of contrast results (from emmeans)}

\item{xlab}{Character string to label x-axis. Defaults to \code{treatment}}

\item{ylab}{Character string to label y-axis. Defaults to \code{variable}}
}
\description{
Create Boxplots with Group Means ± CI and Significance Bars
}
