% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_cross_fostered.R
\name{exclude_cross_fostered}
\alias{exclude_cross_fostered}
\title{Exclude Cross-Fostered Animals}
\usage{
exclude_cross_fostered(
  dataset,
  subj_id = "Animal ID",
  cross_fostered = c("9_31_1", "9_31_2", "9_31_3", "9_31_4")
)
}
\arguments{
\item{dataset}{Data frame of the data}

\item{subj_id}{Character identifying the subject ID variable}

\item{cross_fostered}{Vector of cross-fostered animals}
}
\description{
Exclude Cross-Fostered Animals
}
