% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_analysis.R
\name{longitudinal_analysis}
\alias{longitudinal_analysis}
\title{Analyze Longitudinal MAR data}
\usage{
longitudinal_analysis(dataset, outcome)
}
\arguments{
\item{dataset}{A \code{\link[=data.frame]{data.frame()}} containing the outcome of interest in long format}

\item{outcome}{Name of the outcome variable}
}
\value{
What does this return?
}
\description{
Analyze Longitudinal MAR data
}
