% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_helpers.R
\name{make_posthoc_flextable}
\alias{make_posthoc_flextable}
\title{Posthoc Test Helper Function}
\usage{
make_posthoc_flextable(df, metric = "Mean_Diff")
}
\arguments{
\item{df}{Data frame of the data}

\item{metric}{The column name and header label you want to use}
}
\value{
Return the table of posthoc estimation table
}
\description{
Posthoc Test Helper Function
}
