% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_helpers.R
\name{prop_ci}
\alias{prop_ci}
\title{Posthoc Test Helper Function}
\usage{
prop_ci(x1, n1, x2, n2, conf.level = 0.95)
}
\arguments{
\item{dataset}{Data frame of the data}

\item{var}{Response variable}

\item{control_group}{Level of the control group}
}
\value{
Return the table of posthoc estimation table
}
\description{
Posthoc Test Helper Function
}
