% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_time_to_event.R
\name{summarize_time_to_event}
\alias{summarize_time_to_event}
\title{Summarize the Time-to-event results}
\usage{
summarize_time_to_event(result_list)
}
\arguments{
\item{result_list}{A list containing the time-to-event results for each
measurement at each PND}
}
\description{
Summarize the Time-to-event results
}
