% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_event.R
\name{time_to_event}
\alias{time_to_event}
\title{Developmental Milestons Time-to-event analysis}
\usage{
time_to_event(
  dataset,
  var_list = c("Negative Geotaxis", "Cliff Avoidance", "Circle Traverse")
)
}
\arguments{
\item{dataset}{Data.frame}

\item{var_list}{Vector of milestone variables with censoring.
Default = c("Negative Geotaxis", "Cliff Avoidance", "Circle Traverse")}
}
\description{
Developmental Milestons Time-to-event analysis
}
