% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_helpers.R
\name{run_posthoc_test_cat_h1}
\alias{run_posthoc_test_cat_h1}
\title{Posthoc Test Helper Function}
\usage{
run_posthoc_test_cat_h1(data, var, control_group = "Adjuvant+Saline")
}
\arguments{
\item{var}{Response variable}

\item{control_group}{Level of the control group}

\item{dataset}{Data frame of the data}
}
\value{
Return the table of posthoc estimation table
}
\description{
Posthoc Test Helper Function
}
