% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_helpers.R
\name{run_posthoc_test_num_h2}
\alias{run_posthoc_test_num_h2}
\title{Posthoc Test Helper Function}
\usage{
run_posthoc_test_num_h2(data, vars, exclude_control = "Adjuvant+Saline")
}
\arguments{
\item{dataset}{Data frame of the data}

\item{var}{Response variable}

\item{control_group}{Level of the control group}
}
\value{
Return the table of posthoc estimation table
}
\description{
Posthoc Test Helper Function
}
